* commit 'f95bd38265c57ca786f7c133c28fccc8d1706e12':
  Address review comments
  Exclude unsupported unit tests
  Address review comments
  Remove protractor tests
  Revise <sf-form> usage

+++ src/app/lib/pipes/app-json.pipe.ts
@@ -48,7 +48,9 @@ export class AppJsonPipe implements PipeTransform {
       }
       // Remove all custom fields starting with __$ and empty fields.
       else if(this.key?.startsWith('__$') || typeof x === 'function' || Util.isEmpty(x)) {
-        this.delete();
+        if(this.notRoot) {
+          this.delete();
+        }
       }
     });
